# apps/api/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseBoard API",
    "description": "REST API for PulseBoard: authentication (JWT + refresh cookie), poll CRUD, anonymous or authenticated responses, and analytics. WebSocket real-time updates are not described here.\n\n**Auth:** Protected routes expect `Authorization: Bearer <access_token>`. Register/login set an HTTP-only `refreshToken` cookie used by `POST /api/auth/refresh`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token returned from register, login, refresh, or OAuth callback query param."
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": ["error", "details"],
        "properties": {
          "error": { "type": "string", "example": "Validation failed" },
          "details": {
            "type": "object",
            "description": "Zod flatten() output",
            "additionalProperties": true
          }
        }
      },
...
      "PollOptionInput": {
        "type": "object",
        "required": ["text", "order"],
        "properties": {
          "text": { "type": "string", "minLength": 1, "maxLength": 200 },
          "order": { "type": "integer", "minimum": 0 }
        }
      },
      "PollQuestionInput": {
        "type": "object",
        "required": ["text", "order", "options"],
        "properties": {
          "text": { "type": "string", "minLength": 1, "maxLength": 500 },
          "isRequired": { "type": "boolean", "default": true },
          "order": { "type": "integer", "minimum": 0 },
          "options": {
            "type": "array",
            "minItems": 2,
            "maxItems": 10,
            "items": { "$ref": "#/components/schemas/PollOptionInput" }
          }
        }
      },
      "CreatePollRequest": {
        "type": "object",
        "required": ["title", "expiresAt", "questions"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 1000 },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 datetime with offset"
          },
          "isAnonymous": { "type": "boolean", "default": false },
          "showResults": { "type": "boolean", "default": false },
          "questions": {
            "type": "array",
            "minItems": 1,
            "maxItems": 20,
            "items": { "$ref": "#/components/schemas/PollQuestionInput" }
          }
        }
      },
      "UpdatePollRequest": {
        "type": "object",
        "description": "Partial update; questions cannot be changed via this endpoint.",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 1000 },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "isAnonymous": { "type": "boolean" },
          "showResults": { "type": "boolean" }
        }
      },
      "PollOption": {
        "type": "object",
        "required": ["id", "questionId", "text", "order"],
        "properties": {
          "id": { "type": "string" },
          "questionId": { "type": "string" },
          "text": { "type": "string" },
          "order": { "type": "integer" }
        }
      },
      "PollQuestion": {
        "type": "object",
        "required": ["id", "pollId", "text", "isRequired", "order", "options"],
        "properties": {
          "id": { "type": "string" },
          "pollId": { "type": "string" },
          "text": { "type": "string" },
          "isRequired": { "type": "boolean" },
          "order": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "options": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollOption" }
          }
        }
      },
      "Poll": {
        "type": "object",
        "required": [
          "id",
          "title",
          "creatorId",
          "isAnonymous",
          "isPublished",
          "isActive",
          "showResults",
          "expiresAt",
          "createdAt",
          "updatedAt",
          "questions"
        ],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "creatorId": { "type": "string" },
          "isAnonymous": { "type": "boolean" },
          "isPublished": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "showResults": { "type": "boolean" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollQuestion" }
          }
        }
      },
      "PollPublic": {
        "type": "object",
        "description": "Poll resource returned by GET /api/polls/{id}; omits creatorId.",
        "required": [
          "id",
          "title",
          "isAnonymous",
          "isPublished",
          "isActive",
          "showResults",
          "expiresAt",
          "createdAt",
          "updatedAt",
          "questions"
        ],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "isAnonymous": { "type": "boolean" },
          "isPublished": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "showResults": { "type": "boolean" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollQuestion" }
          }
        }
      },
      "PollRow": {
        "type": "object",
        "description": "Poll table row without nested questions (e.g. publish response).",
        "required": [
          "id",
          "title",
          "creatorId",
          "isAnonymous",
          "isPublished",
          "isActive",
          "showResults",
          "expiresAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "creatorId": { "type": "string" },
          "isAnonymous": { "type": "boolean" },
          "isPublished": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "showResults": { "type": "boolean" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PollListItem": {
        "allOf": [
          { "$ref": "#/components/schemas/PollRow" },
          {
            "type": "object",
            "required": ["responseCount"],
            "properties": {
              "responseCount": {
                "type": "integer",
                "minimum": 0,
                "description": "Number of submitted responses"
              }
            }
          }
        ]
      },
      "PollListResponse": {
        "type": "object",
        "required": ["polls"],
        "properties": {
          "polls": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PollListItem" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/polls": {
      "post": {
        "tags": ["Polls"],
        "summary": "Create poll",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePollRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poll created with nested questions and options",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Poll" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorMessage" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Polls"],
        "summary": "List my polls",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Polls for the authenticated user with response counts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollListResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorMessage" }
              }
            }
          }
        }
      }
    },
    "/api/polls/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" },
          "description": "Poll ID (cuid)"
        }
      ],
      "get": {
        "tags": ["Polls"],
        "summary": "Get poll (public)",
        "description": "Returns poll with questions; omits `creatorId`. May flip `isActive` if expired.",
        "responses": {
          "200": {
            "description": "Poll",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollPublic" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": { "error": "Resource not found" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Polls"],
        "summary": "Update poll metadata",
        "description": "Owner only. Does not modify questions.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePollRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated poll with nested relations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Poll" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorMessage" }
              }
            }
          },
          "403": {
            "description": "Not the owner",
            "content": {
              "application/json": {
                "example": { "error": "Access denied" }
              }
            }
          },
          "404": {
            "description": "Poll not found",
            "content": {
              "application/json": {
                "example": { "error": "Resource not found" }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Polls"],
        "summary": "Delete poll",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorMessage" }
              }
            }
          },
          "403": {
            "description": "Not the owner",
            "content": {
              "application/json": {
                "example": { "error": "Access denied" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": { "error": "Resource not found" }
              }
            }
          }
        }
      }
    },
    "/api/polls/{id}/publish": {
      "post": {
        "tags": ["Polls"],
        "summary": "Publish poll",
        "description": "Owner only. Sets `isPublished` to true. Response body is the poll row only (no nested questions).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PollRow" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorMessage" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": { "error": "Access denied" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": { "error": "Resource not found" }
              }
            }
          }
        }
      }
    },
...
  }
}